i18n: handle strings in kdevtemplate files
The 'qml-plasmoid' is excluded on purpose in this case, since the file
has many translations already, and we are too close to the periodic
release date of frameworks: this means extracting it now would remove
all the existing translations, giving a couple of days only to
translators to provide them back.

Once version 5.27.0 is released, this restriction will be lifted, as
there will be plenty of time to translate everything back.