[Containment Interface] Don't add containment actions to applet actions on desktop
For a panel it makes sense to have panel options easily available on every applet
but on the desktop you typically have some free space to access the containment actions.

This is especially confusing if you have a Folder View on your Folder View desktop
and end up with a "Folder View Options" menu that affects the desktop behind it rather than the applet.

Differential Revision: https://phabricator.kde.org/D8009