Don't set a window icon in Plasma::Dialog
Summary:
Setting a window icon is costly enough to be worth avoiding when it's
not actually needed.

This is a resurrection of David's old patch:

https://git.reviewboard.kde.org/r/128484/

The concerns in the old discussion that led to its rejection are no
longer valid today: In the meantime we implemented a way for apps to
announce their .desktop file in a window hint, which KWin (and libtm)
will fall back to to look up an icon if not set. As plasmashell does
this, we can drop the setIcon call here and won't regress.

Test Plan:
The pinned systray popup dialog still gets the Plasma icon in the
Present Windows effect.

Reviewers: #plasma, #frameworks, davidedmundson, graesslin

Subscribers: plasma-devel

Tags: #plasma, #frameworks

Differential Revision: https://phabricator.kde.org/D9209