Fix fading node when one textured is atlassed.
source and target were swapped in both the QSGNode::updateState/bind
and in the fragment shader.

When Atlas support was added, the texture position was put the right way
round; which made things backwards so source/target also got swapped
instead of fixed properly.

But this leads a glitch if one texture was in the atlas and one
wasn't (not common)

This puts everything the right way round again.

Reviewed-by: Kai Broulik