[DialogShadows] Use 0 offset for disabled borders on Wayland
Summary:
On Wayland, DialogShadows doesn't attach empty tiles for disabled
borders. Yet, it sets offset of 1 for them. This results in visual
artifacts like this

{F6206088, layout=center, size=full}

//KWin thinks that there is 1px tall shadow in the bottom.//

In general we don't need to pass empty shadow tiles for disabled
borders. Thus, that's fine to set 0 offset for them.

Test Plan:
* Started kwin_wayland:

    dbus-launch kwin_wayland --windowed --xwayland

* Launched latte-dock and kate:

    unset DISPLAY
    QT_QPA_PLATFORM=wayland latte-dock &
    QT_QPA_PLATFORM=wayland kate

* Opened calendar popup:

{F6206090, layout=center, size=full}

Reviewers: davidedmundson

Reviewed By: davidedmundson

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D14946