Air/Oxygen desktoptheme: fix progressbar height using "hint-bar-size"
Summary:
During Pre-Plasma5 the hard-coded height of 20 of the progressbar was
changed to instead be taken from the theme via the "hint-bar-size":
commit cee9c928e3855b68daf14760ed71aa961f231172

As fallback for old themes in absence of the hint the default height
since some time later has been derived from the sum of the heights of
the elements "bar-inactive-top" & "bar-inactive-bottom".

Which though for the Air and Oxygen themes results in the small heights
of respective 10 & 4. As a result the visual appearance of the theme
differs from what people remember of the time when the themes were
designed.
Ideally the theme designer would tune the actual value to something they
prefer. Until then setting the hint to restore the visual appearance over
the currently pretty slim or even too slim one should improve things
already.

Reviewers: #plasma, #vdg

Subscribers: broulik, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D20602