Allow applets/containments/wallpaper to defer UIReadyConstraint
Summary:
This is needed so that we don't remove the splash till the wallpaper is
loaded, but also allows doing this in a generic way.

Version 2 of the patch using a normal property.

We need to have the virtual so that the containment waits for the
wallpaper.

Test Plan:
Used with wallpaper patch
Added debug to ensure it happened after the stackview in the wallpaper
is pushed

Reviewers: #plasma

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D22852