make button icons follow a reasonable size and layout
Summary:
with the Plasma style buttons which have an icon look wrong. The icon size is too big and in consecuence the whole button is drawn way too big.
With this patch the icon is set to a standard size and drawn in the button layout, in a way it looks consistent with other button with no icons.

example:

Before path:

{F7270612}

{F7270611}

With patch:

{F7270610}

{F7270609}

BUG: 399952
FIXED-IN: 5.62

Reviewers: #plasma, mart, ngraham

Reviewed By: ngraham

Subscribers: ngraham, kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D23352