[PC3/button] Elide text always
Summary:
We always want the label to have the width managed by the layout to fill
all available space.

Test Plan:
In tests/components/button3.qml the test

"long text but constrained, should be 150px and elided" now is correctly
elided rather than the text spewing past the end of the button. The text
does not overlap the icon.

Reviewers: #plasma, apol, ngraham

Reviewed By: apol, ngraham

Subscribers: kde-frameworks-devel

Tags: #frameworks

Differential Revision: https://phabricator.kde.org/D24909