[Plasma Dialog] Use QXcbWindowFunctions for setting window types Qt WindowFlags doesn't know
... but Qt XCB does, such as Notification type.
This ensures QXcbWindow keeps the correct window types around so we don't have Qt setting it back
to e.g. Dialog when we change some flags and then having to manually set it back using KWindowSystem.

Only for types that aren't in NET WM like OSD and Critical Notifications will it have to fall back to
the old codepath. It also just uses it directly on Wayland.

Differential Revision: https://phabricator.kde.org/D26000