Remove the panel tooltip icon
Summary:
As discussed in T12778, the icon can never be in a correct state: either it's
redundant with the icon you're hovering over that made the tooltip appear, or
else it's different from the icon you're hovering over, which is inconsistent.

Since either case is undesirable, let's just remove the icon from the tooltip
entirely. This reduces the visual noise in the tooltip and solves the entire
class of bugs automatically.

Test Plan:
{F8150019}
{F8150017}
{F8150021}

Reviewers: #vdg, #plasma, cblack, niccolove, apol

Reviewed By: #vdg, cblack, niccolove, apol

Subscribers: cblack, kde-frameworks-devel

Tags: #frameworks

Maniphest Tasks: T12778

Differential Revision: https://phabricator.kde.org/D27848