[calendar] Reduce month label's size
Level 1 is too large for this, as it's typically used for the entire
window or popup's title. The smaller level 2 is therefore more
appropriate so that when the parent window supplies its own level 1
heading, the calendar month label don't inappropriately be the same size.

Needed for https://invent.kde.org/plasma/plasma-workspace/-/merge_requests/98